requirements.txt
pytest==8.3.4
selene==2.0.0rc9
selenium~=4.23.1
python-dotenv==1.0.1
allure-pytest==2.13.5